Fix method descriptors for Python 3.8
Importing leveldb in Python 3.8 raises `SystemError: bad call flags`.
This is due to a change in Python where the method descriptor flags are
checked at import time, rather than when the methods are called.  This
change fixes the incorrect method description.

Adapted from @vstinner's fix for the same problem in libcomps.

https://github.com/rpm-software-management/libcomps/pull/50